Metody serwisu

+++ src/main/java/com/hotel/repositories/RoomRepository.java
@@ -1,8 +1,11 @@
 package com.hotel.repositories;
 
 import com.hotel.entity.Room;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoomRepository extends CrudRepository<Room, Integer> {
+import java.util.List;
+
+public interface RoomRepository extends CrudRepository<Room, Long> {
+//    @Override
+//    List<Room> findAll();
 }
